Split hosted public join launcher test
